# src/lambda/binService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de rutas y emergencias",
        "title": "IntegracionDeAplicaciones2 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/tachos": {
            "get": {
                "description": "Devuelve todos los tachos con neighborhood, latitud, longitud, estado y capacidad",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tachos"
                ],
                "summary": "Obtener todos los tachos",
                "responses": {
                    "200": {
                        "description": "Lista de todos los tachos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.TachoCompleto"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un tacho guardándolo tanto en MySQL como en Neo4j",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tachos"
                ],
                "summary": "Crear un nuevo tacho",
                "parameters": [
                    {
                        "description": "Datos del tacho a crear",
                        "name": "tacho",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateTachoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tacho creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/services.CreateTachoResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un tacho tanto de MySQL como de MongoDB usando el ID de MongoDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tachos"
                ],
                "summary": "Eliminar un tacho",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del tacho en MongoDB",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tacho eliminado exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Tacho no encontrado en ninguna base",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services.CreateTachoRequest": {
            "type": "object",
            "properties": {
                "capacidad": {
                    "type": "number"
                },
                "id": {
                    "description": "Datos para MongoDB (id, lat, lon, neighborhood)",
                    "type": "integer"
                },
                "id_estado": {
                    "type": "integer"
                },
                "id_tipo": {
                    "description": "Datos para MySQL",
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "neighborhood": {
                    "description": "Número de barrio",
                    "type": "integer"
                }
            }
        },
        "services.CreateTachoResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "mongo_id": {
                    "type": "integer"
                },
                "tacho_id": {
                    "type": "integer"
                }
            }
        },
...
        "services.TachoCompleto": {
            "type": "object",
            "properties": {
                "capacidad": {
                    "type": "number"
                },
                "estado": {
                    "type": "string"
                },
                "id_tacho": {
                    "type": "integer"
                },
                "latitud": {
                    "type": "number"
                },
                "longitud": {
                    "type": "number"
                },
                "neighborhood": {
                    "type": "integer"
                }
            }
        }
    }
}